Clean up probability models (#195, #197).
