+++ maxodiff-core/src/main/java/org/p2gx/maxodiff/core/analysis/ValidationModel.java
@@ -1,10 +1,9 @@
 package org.p2gx.maxodiff.core.analysis;
 
 
+import org.p2gx.maxodiff.core.analysis.impl.ValidationWeightedRankDiffImpl;
 import org.p2gx.maxodiff.core.model.DifferentialDiagnosis;
 import org.p2gx.maxodiff.core.analysis.impl.ValidationRankDiffImpl;
-import org.p2gx.maxodiff.core.analysis.impl.ValidationScoreDiffImpl;
-import org.p2gx.maxodiff.core.analysis.impl.ValidationWeightedRankDiffImpl;
 
 
 import java.util.List;
@@ -21,11 +20,6 @@ public interface ValidationModel {
     static ValidationModel weightedRankDiff(List<DifferentialDiagnosis> initialDifferentialDiagnoses,
                                             List<DifferentialDiagnosis> maxoDifferentialDiagnoses) {
         return new ValidationWeightedRankDiffImpl(initialDifferentialDiagnoses, maxoDifferentialDiagnoses);
-    }
-
-    static ValidationModel scoreDiff(List<DifferentialDiagnosis> initialDifferentialDiagnoses,
-                                     List<DifferentialDiagnosis> maxoDifferentialDiagnoses) {
-        return new ValidationScoreDiffImpl(initialDifferentialDiagnoses, maxoDifferentialDiagnoses);
     }
 
 }
