--- maxodiff-core/src/main/java/org/p2gx/maxodiff/core/analysis/impl/DiseaseSoftmaxProbabilityImpl.java
@@ -1,28 +1,0 @@
-package org.p2gx.maxodiff.core.analysis.impl;
-
-import org.p2gx.maxodiff.core.model.DiseaseModelProbability;
-import org.p2gx.maxodiff.core.model.DifferentialDiagnosis;
-import org.monarchinitiative.phenol.ontology.data.TermId;
-
-import java.util.List;
-
-/**
- * The  softmax function is a normalized exponential function that converts scores into probabilities.
- */
-public final class DiseaseSoftmaxProbabilityImpl extends DiseaseModelProbabilityImpl  implements DiseaseModelProbability {
-
-    private final double ddScoreSum;
-
-    public DiseaseSoftmaxProbabilityImpl(List<DifferentialDiagnosis> differentialDiagnoses) {
-        super(differentialDiagnoses);
-        ddScoreSum = differentialDiagnoses.stream()
-                .mapToDouble(dd -> Math.exp(dd.score()))
-                .sum();
-    }
-
-    public double probability(TermId targetDiseaseId) {
-        double targetDiagnosisScore = getTargetDiseaseDiagnosis(targetDiseaseId).score();
-        return Math.exp(targetDiagnosisScore) / ddScoreSum;
-    }
-
-}
