+++ maxodiff-core/src/main/java/org/p2gx/maxodiff/core/analysis/impl/ValidationRankDiffImpl.java
@@ -12,13 +12,13 @@ import java.util.Map;
  * Validate the effectiveness of the maxodiff algorithm using the sum of differences of disease ranks before and after
  * applying the maxodiff algorithm.
  */
-public final class ValidationRankDiffImpl extends ValidationModelImpl  implements ValidationModel {
+public final class ValidationRankDiffImpl implements ValidationModel {
 
+    /** A map of MAxO term ids : differential diagnoses after running the maxodiff analysis **/
     private final Map<TermId, Double> rankChanges = new HashMap<>();
 
     public ValidationRankDiffImpl(List<DifferentialDiagnosis> initialDifferentialDiagnoses,
                                   List<DifferentialDiagnosis> maxoDifferentialDiagnoses) {
-        super(initialDifferentialDiagnoses, maxoDifferentialDiagnoses);
 
         for (DifferentialDiagnosis initialDiagnosis : initialDifferentialDiagnoses) {
             TermId termId = initialDiagnosis.diseaseId();
