+++ maxodiff-core/src/main/java/org/p2gx/maxodiff/core/analysis/refinement/DiffDiagRefinerImpl.java
@@ -4,7 +4,6 @@ package org.p2gx.maxodiff.core.analysis.refinement;
 import org.monarchinitiative.phenol.ontology.data.MinimalOntology;
 import org.p2gx.maxodiff.core.analysis.MySimpleTerm;
 import org.p2gx.maxodiff.core.analysis.RankedMaxoResult;
-import org.p2gx.maxodiff.core.analysis.SimpleTerm;
 import org.p2gx.maxodiff.core.diffdg.DDxEngine;
 
 import org.p2gx.maxodiff.core.io.MdContext;
@@ -78,12 +77,10 @@ public class DiffDiagRefinerImpl implements DiffDiagRefiner {
                                 Map<TermId, Set<TermId>> maxoToHpoTermIdMap,
                                 List<HpoFrequency> hpoFrequenciesNDiseases) {
 
-        DiseaseModelProbability diseaseModelProbability = DiseaseModelProbability.ranked(initialDiagnoses);
 
         MaxoHpoTermProbabilities maxoHpoTermProbabilities = new MaxoHpoTermProbabilities(hpoDiseases,
                 hpoToMaxoTermMap,
-                initialDiagnoses,
-                diseaseModelProbability);
+                initialDiagnoses);
 
 
         return new RankMaxo(maxoToHpoTermIdMap, maxoHpoTermProbabilities, engine,
