+++ maxodiff-core/src/main/java/org/p2gx/maxodiff/core/io/MdResources.java
@@ -2,9 +2,7 @@ package org.p2gx.maxodiff.core.io;
 
 import org.monarchinitiative.phenol.annotations.formats.hpo.HpoDiseases;
 import org.monarchinitiative.phenol.ontology.data.MinimalOntology;
-import org.p2gx.maxodiff.core.analysis.HTMLFrequencyMap;
 import org.p2gx.maxodiff.core.analysis.MySimpleTerm;
-import org.p2gx.maxodiff.core.analysis.SimpleTerm;
 import org.p2gx.maxodiff.core.phenomizer.IcMicaData;
 
 import java.util.Map;
