+++ maxodiff-core/src/main/java/org/p2gx/maxodiff/core/model/DiseaseModelProbability.java
@@ -1,8 +1,6 @@
 package org.p2gx.maxodiff.core.model;
 
-import org.p2gx.maxodiff.core.analysis.impl.DiseaseExponentialDecayProbabilityImpl;
 import org.p2gx.maxodiff.core.analysis.impl.DiseaseRankedProbabilityImpl;
-import org.p2gx.maxodiff.core.analysis.impl.DiseaseSoftmaxProbabilityImpl;
 import org.monarchinitiative.phenol.ontology.data.TermId;
 
 import java.util.List;
@@ -15,18 +13,6 @@ public interface DiseaseModelProbability {
 
     static DiseaseModelProbability ranked(List<DifferentialDiagnosis> diagnoses) {
         return new DiseaseRankedProbabilityImpl(diagnoses);
-    }
-
-    static DiseaseModelProbability softmax(List<DifferentialDiagnosis> diagnoses) {
-        return new DiseaseSoftmaxProbabilityImpl(diagnoses);
-    }
-
-    static DiseaseModelProbability exponentialDecay(List<DifferentialDiagnosis> diagnoses) {
-        return new DiseaseExponentialDecayProbabilityImpl(diagnoses, DEFAULT_LAMBDA);
-    }
-
-    static DiseaseModelProbability exponentialDecay(List<DifferentialDiagnosis> diagnoses, double lambda) {
-        return new DiseaseExponentialDecayProbabilityImpl(diagnoses, lambda);
     }
 
 }
