+++ maxodiff-core/src/main/java/org/p2gx/maxodiff/core/service/BiometadataServiceImpl.java
@@ -1,7 +1,6 @@
 package org.p2gx.maxodiff.core.service;
 
 import org.p2gx.maxodiff.core.analysis.MySimpleTerm;
-import org.p2gx.maxodiff.core.analysis.SimpleTerm;
 import org.monarchinitiative.phenol.annotations.formats.hpo.HpoDisease;
 import org.monarchinitiative.phenol.annotations.formats.hpo.HpoDiseases;
 import org.monarchinitiative.phenol.ontology.data.MinimalOntology;
