+++ maxodiff-core/src/test/java/org/p2gx/maxodiff/core/analysis/DiseaseModelProbabilityTest.java
@@ -36,9 +36,6 @@ public class DiseaseModelProbabilityTest {
 
     public sealed interface TestOutcome {
         record OkRanked(double diseaseProbability) implements TestOutcome {}
-        record OkSoftmax(double diseaseProbability) implements TestOutcome {}
-        record OkExpDecay(double diseaseProbability) implements TestOutcome {}
-        record OkExpDecay1(double diseaseProbability) implements TestOutcome {}
         record Error(Supplier<? extends PhenolRuntimeException> exceptionSupplier) implements TestOutcome {}
     }
 
@@ -54,15 +51,6 @@ public class DiseaseModelProbabilityTest {
                 new TestIndividual("ranked",
                         TARGET_ID,
                         new TestOutcome.OkRanked(0.09)),
-                new TestIndividual("softmax",
-                        TARGET_ID,
-                        new TestOutcome.OkSoftmax(0.072)),
-                new TestIndividual("exponential decay",
-                        TARGET_ID,
-                        new TestOutcome.OkExpDecay(0.632)),
-                new TestIndividual("exponential decay",
-                        TARGET_ID,
-                        new TestOutcome.OkExpDecay1(0.393)),
                 new TestIndividual("no disease",
                         TermId.of("OMIM:123456"),
                         new TestOutcome.Error(() ->
@@ -77,18 +65,6 @@ public class DiseaseModelProbabilityTest {
         switch (testCase.expectedOutcome()) {
             case TestOutcome.OkRanked(double expectedResult) ->
                     assertEquals(expectedResult, DiseaseModelProbability.ranked(DIFFERENTIAL_DIAGNOSES).probability(targetId),
-                            1.e-3,
-                            "Incorrect evaluation for: " + testCase.description());
-            case TestOutcome.OkSoftmax(double expectedResult) ->
-                    assertEquals(expectedResult, DiseaseModelProbability.softmax(DIFFERENTIAL_DIAGNOSES).probability(targetId),
-                            1.e-3,
-                            "Incorrect evaluation for: " + testCase.description());
-            case TestOutcome.OkExpDecay(double expectedResult) ->
-                    assertEquals(expectedResult, DiseaseModelProbability.exponentialDecay(DIFFERENTIAL_DIAGNOSES).probability(targetId),
-                            1.e-3,
-                            "Incorrect evaluation for: " + testCase.description());
-            case TestOutcome.OkExpDecay1(double expectedResult) ->
-                    assertEquals(expectedResult, DiseaseModelProbability.exponentialDecay(DIFFERENTIAL_DIAGNOSES, 0.5).probability(targetId),
                             1.e-3,
                             "Incorrect evaluation for: " + testCase.description());
             case TestOutcome.Error(Supplier<? extends RuntimeException> exceptionSupplier) ->
