+++ maxodiff-core/src/test/java/org/p2gx/maxodiff/core/analysis/ValidationModelTest.java
@@ -1,33 +1,17 @@
 package org.p2gx.maxodiff.core.analysis;
 
 import org.junit.jupiter.api.Test;
-
 import org.p2gx.maxodiff.core.TestResources;
-import org.monarchinitiative.phenol.annotations.formats.hpo.HpoDiseases;
 import org.p2gx.maxodiff.core.model.DifferentialDiagnosis;
-import org.p2gx.maxodiff.core.model.DiseaseModelProbability;
-import org.p2gx.maxodiff.core.model.MaxoHpoTermProbabilities;
 
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
 public class ValidationModelTest {
 
-    private final static HpoDiseases hpoDiseases = TestResources.hpoDiseases();
     private final static List<DifferentialDiagnosis> initialDiagnoses = TestResources.getExampleDiagnoses().stream().toList();
-    private final static Map<MySimpleTerm, Set<MySimpleTerm>> hpoToMaxoTermMap = TestResources.hpoToMaxo();
-
-    private final static Map<String, Set<String>> maxoToHpoTermIdMap = MaxoHpoTermIdMaps.getMaxoToHpoTermIdMap(hpoToMaxoTermMap);
-    private final static MaxoHpoTermProbabilities maxoHpoTermProbabilities =
-            new MaxoHpoTermProbabilities(hpoDiseases,
-                                         hpoToMaxoTermMap,
-                                         initialDiagnoses,
-                                         DiseaseModelProbability.ranked(initialDiagnoses));
-
     private final static List<DifferentialDiagnosis> maxoDiagnoses = TestResources.getExampleMaxoDiagnoses().stream().toList();
 
 
@@ -49,13 +33,5 @@ public class ValidationModelTest {
         assertEquals(1.005, validationScore, 1e-3);
     }
 
-    /**
-     * This tests validating the maxodiff analysis using the sum of differences in disease scores.
-     */
-    @Test
-    public void testScoreDiffValidationModel() {
-        double validationScore = ValidationModel.scoreDiff(initialDiagnoses, maxoDiagnoses).validationScore();
-        assertEquals(8.487, validationScore, 1e-3);
-    }
 
 }
