+++ maxodiff-html-results/src/main/java/org/p2gx/maxodiff/html/results/maxoHpo/MdHtmlResult.java
@@ -1,6 +1,5 @@
 package org.p2gx.maxodiff.html.results.maxoHpo;
 
-import org.monarchinitiative.phenol.ontology.data.TermId;
 import org.p2gx.maxodiff.core.analysis.*;
 
 import java.util.*;
